Validation + message erreur admin cours

+++ Admin/GérerLesCours.aspx.vb
@@ -43,32 +43,57 @@ Partial Class Admin_GérerLesCours
         lecontext.AddObject("CoursSet", leCoursAjouté)
         lecontext.SaveChanges()
         hFieldNoCours.Value = leCoursAjouté.noCours
+        lViewModifierCours.EditIndex = 0
         mViewCours.ActiveViewIndex = 1
     End Sub
 
     Protected Sub lViewCours_ItemCommand(ByVal sender As Object, ByVal e As System.Web.UI.WebControls.ListViewCommandEventArgs) Handles lViewCours.ItemCommand
         Dim leNoCours As String = e.CommandArgument
         If e.CommandName = "Delete" Then
-            Dim leCoursADeleter As Cours = (From monCours In lecontext.CoursSet Where monCours.noCours = leNoCours Select monCours).First
-            lecontext.CoursSet.DeleteObject(leCoursADeleter)
-            lecontext.SaveChanges()
-            Response.Redirect("~/Admin/GérerLesCours.aspx")
+            Dim peutDeleter As Boolean = False
+            For Each leCours In lecontext.CoursSet
+                If leCours.noCours = leNoCours Then
+                    peutDeleter = True
+                End If
+            Next
+            If peutDeleter Then
+                Dim leCoursADeleter As Cours = (From monCours In lecontext.CoursSet Where monCours.noCours = leNoCours Select monCours).First
+                lecontext.CoursSet.DeleteObject(leCoursADeleter)
+                lecontext.SaveChanges()
+                'lblMessage.Text = "Le cours a bien été supprimé"
+                Response.Redirect("~/Admin/GérerLesCours.aspx")
+            Else
+                lblMessage.Text = "Le Cours n'existe pas dans la base de données"
+            End If
         ElseIf e.CommandName = "Modifier" Then
             hFieldNoCours.Value = e.CommandArgument
             lViewModifierCours.EditIndex = 0
             mViewCours.ActiveViewIndex = 1
+            lblMessage.Text = ""
         End If
 
     End Sub
 
     Protected Sub lViewModifierCours_ItemCommand(ByVal sender As Object, ByVal e As System.Web.UI.WebControls.ListViewCommandEventArgs) Handles lViewModifierCours.ItemCommand
         If e.CommandName = "Annuler" Then
-            Dim leNoCours As String = hFieldNoCours.Value
-            Dim leCoursADeleter As Cours = (From monCours In lecontext.CoursSet Where monCours.noCours = leNoCours Select monCours).First
-            If leCoursADeleter.Nom = "Entrez un nom" Then
-                lecontext.CoursSet.DeleteObject(leCoursADeleter)
-                lecontext.SaveChanges()
-                mViewCours.ActiveViewIndex = 0
+            Dim peutDeleter As Boolean = False
+            For Each leCours In lecontext.CoursSet
+                If leCours.noCours = hFieldNoCours.Value Then
+                    peutDeleter = True
+                End If
+            Next
+            If peutDeleter Then
+                Dim leNoCours As String = hFieldNoCours.Value
+                Dim leCoursADeleter As Cours = (From monCours In lecontext.CoursSet Where monCours.noCours = leNoCours Select monCours).First
+                If leCoursADeleter.Nom = "Entrez un nom" Then
+                    lecontext.CoursSet.DeleteObject(leCoursADeleter)
+                    lecontext.SaveChanges()
+                    mViewCours.ActiveViewIndex = 0
+                    lblMessage.Text = "L'ajout a été annulé"
+                End If
+                lblMessage.Text = "La modification a été annulée"
+            Else
+                lblMessage.Text = "Le Cours n'existe pas dans la base de données"
             End If
         End If
         mViewCours.ActiveViewIndex = 0
@@ -77,6 +102,7 @@ Partial Class Admin_GérerLesCours
     Protected Sub lViewModifierCours_ItemUpdated(ByVal sender As Object, ByVal e As System.Web.UI.WebControls.ListViewUpdatedEventArgs) Handles lViewModifierCours.ItemUpdated
         mViewCours.ActiveViewIndex = 0
         lViewCours.DataBind()
+        lblMessage.Text = "Le cours a été modifié"
     End Sub
 
 End Class
